Add compose function, and HelixSpec.extend
This update adds a new compose function, that allows you to combine
Specs together (as well as regular objects) into a new HelixSpec.

This is powered by the new HelixSpec.extend method, which uses Object.assign
to extend the spec "shape".

The new composed Spec is generated with the usual `.generate` method.