Seeding: Generating different + consistent values for multiple fixtures (#22)
This update fixes the `generate` method for `HelixSpec` which ensures that if
multiple fixtures are being generated, then all of those fixtures use a
different seed index.

This allows for multiple fixtures to be consistently seeded, but have different
seeded results.

The generate accepts the initial seed value (e.g. 2), and increments it for
every iteration that it generates (e.g. 2, 3, 4, 5).